Moved code from Enemy to SkeletonController
This code was in practice just specific to the skeleton enemy, and not enemies in general (like the unigolem).

Also:
* Moved and renamed `Enemy_PlayerDetector` -> `Enemies/Skeleton/Skeleton_PlayerDetector`, as the logic is currently specific to the skeleton enemy,
  and other enemies will not necessarily be interested in trying to detect players
* Renamed `SkeletonController.DealDamage()` -> `OnDealDamage()`
* Renamed `HealthLogic.DealDamage()` -> `OnReceiveDamage()`
