+++ Assets/Prefabs/Enemy/Enemies/Skeleton.prefab
@@ -615,7 +615,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7418487882725981173}
-  - component: {fileID: 6287395116763721580}
+  - component: {fileID: 8622069783345018454}
   - component: {fileID: 696503801798822935}
   m_Layer: 0
   m_Name: PlayerDetector
@@ -638,7 +638,7 @@ Transform:
   m_Father: {fileID: 6262807936475223192}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &6287395116763721580
+--- !u!114 &8622069783345018454
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -650,7 +650,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e28107225c642504e95ee6f0e3939d1d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  enemy: {fileID: 6262807936475223191}
+  skeleton: {fileID: 6262807936475223191}
 --- !u!135 &696503801798822935
 SphereCollider:
   m_ObjectHideFlags: 0
@@ -1971,16 +1971,15 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _currentTarget: {fileID: 0}
-  baseTransform: {fileID: 0}
   currentState: 0
   anim: {fileID: 6942039963572180435}
+  attackDamage: 1
+  scoreValue: 10
   walkSpeed: 1.3
   chaseSpeed: 3
   baseAttackDistance: 2
   playerAttackDistance: 2.5
   playerChaseStopDistance: 8
-  attackDamage: 1
-  scoreValue: 10
 --- !u!114 &6262807936475223195
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -2026,6 +2025,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   lootPrefab: {fileID: 5246800848510687331, guid: b948d203c55455f4288c07c469eda34a, type: 3}
   spawnHeight: 1
+  dropChance: 10
 --- !u!195 &6262807936475223190
 NavMeshAgent:
   m_ObjectHideFlags: 0
